backend/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-RESTful==0.3.10
Flask-Cors==4.0.1
psycopg2-binary==2.9.9
PyMySQL==1.1.1
